# backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API de Gestion des Recettes",
        "description": "API pour gérer les utilisateurs, recettes, catégories, ingrédients et favoris.",
        "contact": {
            "email": "support@votreapp.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/recipes": {
            "get": {
                "tags": [
                    "Recipes"
                ],
                "summary": "Get a list of all recipes",
                "description": "Returns a list of all recipes in the system.",
                "operationId": "d4042afb2fbbddc6df47e660bab9cea7",
                "responses": {
                    "200": {
                        "description": "A list of recipes",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Recipe"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Recipes"
                ],
                "summary": "Create a new recipe",
                "description": "Creates a new recipe with the provided data.",
                "operationId": "b8e5dc90090b088f52bf6ba8ada1a43f",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "title",
                                    "description",
                                    "instructions",
                                    "user_id",
                                    "category_id"
                                ],
                                "properties": {
                                    "title": {
                                        "type": "string",
                                        "example": "Delicious Pancakes"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "A simple and quick pancake recipe."
                                    },
                                    "instructions": {
                                        "type": "string",
                                        "example": "1. Mix ingredients... 2. Cook on medium heat..."
                                    },
                                    "user_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "category_id": {
                                        "type": "integer",
                                        "example": 2
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Recipe created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Recipe"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "The given data was invalid."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/recipes/{id}": {
            "get": {
                "tags": [
                    "Recipes"
                ],
                "summary": "Get recipe by ID",
                "description": "Returns a recipe by its ID.",
                "operationId": "1574c2a4ecb3074887fb892b23321b20",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the recipe",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Recipe data",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Recipe"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Recipe not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Recipe not found"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "Recipes"
                ],
                "summary": "Update recipe by ID",
                "description": "Updates a recipe's details by its ID.",
                "operationId": "83cafcd821c9460d9d673ef89d135e9b",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the recipe to update",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "title": {
                                        "type": "string",
                                        "example": "Updated Pancake Recipe"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Updated description of the recipe."
                                    },
                                    "instructions": {
                                        "type": "string",
                                        "example": "Updated instructions..."
                                    },
                                    "user_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "category_id": {
                                        "type": "integer",
                                        "example": 2
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Recipe updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Recipe"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Recipe not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Recipe not found"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "The given data was invalid."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "Recipes"
                ],
                "summary": "Delete recipe by ID",
                "description": "Deletes a recipe by its ID.",
                "operationId": "9bd6f67654cbd4fbfb963e15d3f0c59a",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the recipe to delete",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Recipe deleted successfully"
                    },
                    "404": {
                        "description": "Recipe not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Recipe not found"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Recipe": {
                "title": "Recipe",
                "description": "Recipe model",
                "properties": {
                    "id": {
                        "title": "ID",
                        "description": "ID of the recipe",
                        "type": "integer",
                        "format": "int64",
                        "example": 1
                    },
                    "title": {
                        "title": "Title",
                        "description": "Title of the recipe",
                        "type": "string",
                        "example": "Delicious Pancakes"
                    },
                    "description": {
                        "title": "Description",
                        "description": "Brief description of the recipe",
                        "type": "string",
                        "example": "A simple and quick pancake recipe."
                    },
                    "instructions": {
                        "title": "Instructions",
                        "description": "Detailed cooking instructions for the recipe",
                        "type": "string",
                        "example": "1. Mix ingredients... 2. Cook on medium heat..."
                    },
                    "user_id": {
                        "title": "User ID",
                        "description": "ID of the user who created the recipe",
                        "type": "integer",
                        "format": "int64",
                        "example": 1
                    },
                    "category_id": {
                        "title": "Category ID",
                        "description": "ID of the category this recipe belongs to",
                        "type": "integer",
                        "format": "int64",
                        "example": 2
                    }
                },
                "type": "object",
                "xml": {
                    "name": "Recipe"
                }
            },
...
        }
    },
...
}